requirements.txt
azure-ai-textanalytics==5.3.0
python-dotenv
azure-ai-language-questionanswering
azure-ai-language-conversations
azure-cognitiveservices-speech
playsound
